Ghosts and Barrels works correctly
Ghosts and Barrels works correctly. The art and coding is in place, and each one of them is working like I set them up.
I need a third type of barrel, which will be golden and spawn bullets from the first and second type combined.
I also need to spawn swords from the top of the screen.

Next thing to do in terms of player features is adding portals. You will be able to create 2 portals. If there is one spawned portal, it will do nothing, which means you need 2 portals to communicate with each other.

Each portal will recognize bullets and the player itself, and send them to the other portal with their current movement applied.