Design change at part 1 of the level
Now there is only 1 cannon with a set of everything (string, wheel and fire). The cannon moves right, but I need to make it fire much more rapidly (every 0.5 second approximately). For this, in CannonSet, I need to set faster animations speed for the cannons, or maybe ditch the animations for a new animation that will simply make the cannon burst, for around 0.7 seconds before firing.