Bat flying animation is ready
Bat flying animation is ready. Now I need to setup 3 points around the hat. 3 bats will move towards them and start turning in circle around the hat, with a changing radius (as I already programmed for the circling stars).

In addition - between phases 2 and 3 the hat will move a bit forward, and I need to work on an interesting way of movement for him.